Move assessmentUtils to utils/assessments folder
Also rename it to utils
