+++ server/utils/assessments/utils.ts
@@ -2,18 +2,18 @@ import { HtmlItem, PageResponse, SummaryListItem, TableRow, Task, TextItem } fro
 import { format, differenceInDays, add } from 'date-fns'
 
 import { ApprovedPremisesAssessment as Assessment, ApprovedPremisesApplication } from '@approved-premises/api'
-import { tierBadge } from './personUtils'
-import { DateFormats } from './dateUtils'
-import { getArrivalDate, getResponseForPage } from './applicationUtils'
-import paths from '../paths/assess'
-import { TasklistPageInterface } from '../form-pages/tasklistPage'
-import Assess from '../form-pages/assess'
-import { UnknownPageError } from './errors'
-import { embeddedSummaryListItem } from './checkYourAnswersUtils'
-import reviewSections from './reviewUtils'
-import Apply from '../form-pages/apply'
-import { kebabCase } from './utils'
-import { documentsFromApplication } from './assessments/documentUtils'
+import { tierBadge } from '../personUtils'
+import { DateFormats } from '../dateUtils'
+import { getArrivalDate, getResponseForPage } from '../applicationUtils'
+import paths from '../../paths/assess'
+import { TasklistPageInterface } from '../../form-pages/tasklistPage'
+import Assess from '../../form-pages/assess'
+import { UnknownPageError } from '../errors'
+import { embeddedSummaryListItem } from '../checkYourAnswersUtils'
+import reviewSections from '../reviewUtils'
+import Apply from '../../form-pages/apply'
+import { kebabCase } from '../utils'
+import { documentsFromApplication } from './documentUtils'
 
 const DUE_DATE_APPROACHING_DAYS_WINDOW = 3
